% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ssc}
\alias{print.ssc}
\title{Print an ssc Object}
\usage{
\method{print}{ssc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ssc"}, as returned by \code{\link{ssc}}.}

\item{...}{Currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Compact one-line summary of a \code{"ssc"} estimation result.
}
